Extract out typed headers and add KeyId
